fix: 최대 현금 자산을 21억 이상으로 수정
기존은 cash 및 모든 자산을 int형으로관리하였으나 long 타입으로 수정했습니다.

+++ src/main/java/com/kakaoasset/portfolio/dto/CashDto.java
@@ -6,9 +6,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor //  cannot deserialize from Object value (no delegate- or property-based Creator)
 public class CashDto {
-    int cash;
+    Long cash;
 
-    public CashDto(int cash) {
+    public CashDto(Long cash) {
         this.cash = cash;
     }
 }
